+++ apps/gpx-web/next.config.mjs
@@ -1,6 +1,8 @@
-/** @type {import('next').NextConfig} */
+/** @type {import("next").NextConfig} */
 const nextConfig = {
   output: "standalone",
+  poweredByHeader: false,
+  compress: true,
 };
 
 export default nextConfig;